Use `git blame` instead of pygit.Repository.blame
pygit.Repository.blame does not support ignoring white spaces when
comparing version of a line between commits. However, when blaming
commits that fixed vulnerabilities in the FFmpeg project, we found
several commits being blamed as suspects when, in actuality, these
commits merely refactored the code. See issue #1 for more information.

The change uses `git blame` instead of pygit.Repository.blame to
workaround the limitation because the limitation is not in pygit but
instead in libgit. Furthermore, there is an open issue from 2015 in the
LibGit2Sharp project (libgit2/libgit2sharp#1177) describing the
limitation.