Parse added, deleted, or context lines only
When parsing a patch produced by git, the lines that matter to
`archeogit blame` are those that were added, deleted, or common
(referred to as context) between two versions of a file. Most patches
only contain these types of lines so there did not seem to be a need to
explicitly restrict parsing of patch. However, a commit in Chromium (See
Issue #2) happened to have a file that we being added without the last
line being terminated with a newline character. The patch caused
`archeogit blame` to raise an unhandled exception causing it to crash.
The change explicitly restricts parsing of lines to those that were
added, deleted, or common between two versions of a file resolving the
issue.

Change resolves #2