Raise an exception of commit is not valid
`archeogit blame` did not check the validity of the commit hash in the
context of the repository before attempting to blame it. As a result, if
the commit hash was invalid, a condition equivalent to a
NullPointerException was reached. The change validates the commit hash
before attempting to blame it.

Change resolves #3