Set `shell` to `True` only on non-Windows platform
According to Python Popen documentation [1], there is no need to set
`shell` to `True` to execute a console-based executable (like `git`).
Although, setting `shell` to `True` does not affect the outcome when the
command being executed completes succesfully. However, if the command
fails, the process intance returned by Popen does not have its
`returncode` set to a non-zero number as one would expect. The issue
described in #4 is essentially a manifestation of this discrepancy
between platforms (Linux and Windows).

Change resolves #4

References

[1] https://docs.python.org/3/library/subprocess.html#subprocess.Popen