Fix bug in path to check when skipping module
The use of submodules adds the directory `ffmpeg` to tests/data but the
directory is empty (i.e. does not have the `ffmpeg` repository cloned).
As a result, checking for the existence of the `ffmpeg` directory will
always return true leading the tests that require the repository to be
cloned to fail.