Fix bug in test case dependent on repository fails
The `test_handler_raises_exception` requires the `ffmpeg` repository to
be cloned into `tests/data/ffmpeg` but if the repository does not exist,
as it would when `archeogit` is cloned without `--recurse-submodules`,
the test fails. The change fixes the bug by skipping the test case if
the repository does not exist.